refactored dom helpers
